server/requirements.txt
firebase-admin
flask
python-dotenv